refactor: Rename and viewmodel setting
