--- app/src/main/java/campus/tech/kakao/map/presentation/viewmodel/MainViewModel.kt
@@ -1,8 +1,0 @@
-package campus.tech.kakao.map.presentation.viewmodel
-
-import android.content.ContentValues
-import androidx.lifecycle.ViewModel
-
-class MainViewModel: ViewModel() {
-
-}
