+++ app/src/main/java/campus/tech/kakao/map/presentation/viewmodel/SearchViewModel.kt
@@ -1,0 +1,17 @@
+package campus.tech.kakao.map.presentation.viewmodel
+
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import campus.tech.kakao.map.data.SearchDbHelper
+import campus.tech.kakao.map.domain.model.SearchData
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+
+class SearchViewModel() : ViewModel() {
+
+}
