+++ app/src/main/res/layout/activity_search.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".presentation.activity.SearchActivity">
+    tools:context=".presentation.view.SearchActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/savedSearchWordRecyclerView"
